refactor(cast): use tempo-contracts is_iso4217_currency and pin tempo rev
Pin tempo to 083592c7. Replace local ISO 4217 reimplementation in cast
with tempo_contracts::precompiles::is_iso4217_currency.

Amp-Thread-ID: https://ampcode.com/threads/T-019cd253-3806-7188-9664-3f300010df92
