--- crates/cast/src/iso4217.rs
@@ -1,39 +1,0 @@
-pub const ISO4217_CODES: &[&str] = &[
-    "AED", "AFN", "ALL", "AMD", "ANG", "AOA", "ARS", "AUD", "AWG", "AZN", "BAM", "BBD", "BDT",
-    "BGN", "BHD", "BIF", "BMD", "BND", "BOB", "BOV", "BRL", "BSD", "BTN", "BWP", "BYN", "BZD",
-    "CAD", "CDF", "CHE", "CHF", "CHW", "CLP", "CLF", "CNY", "COP", "COU", "CRC", "CUP", "CVE",
-    "CZK", "DJF", "DKK", "DOP", "DZD", "EGP", "ERN", "ETB", "EUR", "FJD", "FKP", "GBP", "GEL",
-    "GHS", "GIP", "GMD", "GNF", "GTQ", "GYD", "HKD", "HNL", "HRK", "HTG", "HUF", "IDR", "ILS",
-    "INR", "IQD", "IRR", "ISK", "JMD", "JOD", "JPY", "KES", "KGS", "KHR", "KMF", "KPW", "KRW",
-    "KWD", "KYD", "KZT", "LAK", "LBP", "LKR", "LRD", "LSL", "LYD", "MAD", "MDL", "MGA", "MKD",
-    "MMK", "MNT", "MOP", "MRU", "MUR", "MVR", "MWK", "MXN", "MXV", "MYR", "MZN", "NAD", "NGN",
-    "NIO", "NOK", "NPR", "NZD", "OMR", "PAB", "PEN", "PGK", "PHP", "PKR", "PLN", "PYG", "QAR",
-    "RON", "RSD", "RUB", "RWF", "SAR", "SBD", "SCR", "SDG", "SEK", "SGD", "SHP", "SLE", "SOS",
-    "SRD", "SSP", "STN", "SVC", "SYP", "SZL", "THB", "TJS", "TMT", "TND", "TOP", "TRY", "TTD",
-    "TWD", "TZS", "UAH", "UGX", "USD", "USN", "UYI", "UYU", "UYW", "UZS", "VED", "VES", "VND",
-    "VUV", "WST", "XAF", "XAG", "XAU", "XBA", "XBB", "XBC", "XBD", "XCD", "XDR", "XOF", "XPD",
-    "XPF", "XPT", "XSU", "XTS", "XUA", "XXX", "YER", "ZAR", "ZMW", "ZWL",
-];
-
-pub fn is_valid_iso4217(code: &str) -> bool {
-    let upper = code.to_uppercase();
-    ISO4217_CODES.iter().any(|&c| c == upper)
-}
-
-pub fn iso4217_warning_message(currency: &str) -> String {
-    format!(
-        r#"Warning: "{currency}" is not a recognized ISO 4217 currency code.
-
-If the token you are trying to deploy is a fiat-backed stablecoin, Tempo strongly
-recommends that the currency code field be the ISO-4217 currency code of the fiat
-currency your token tracks (e.g. "USD", "EUR", "GBP").
-
-The currency field is IMMUTABLE after token creation and affects fee payment
-eligibility, DEX routing, and quote token pairing. Only "USD"-denominated tokens
-can be used to pay transaction fees on Tempo.
-
-Learn more:
-  • Tempo TIP-20 docs: https://docs.tempo.xyz/protocol/tip20/overview
-  • ISO 4217 standard: https://www.iso.org/iso-4217-currency-codes.html"#
-    )
-}
